requirements.txt
anyio==3.6.2
certifi @ file:///croot/certifi_1665076670883/work/certifi
click==8.1.3
essentia==2.1b6.dev858
fastapi==0.85.1
Flask==2.2.2
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.3
Werkzeug==2.2.2
zipp==3.9.0
